Put `loading=“lazy”` in the `<img>`.
In short, it should improve performance.
I'll give it a try as I think I can do it while I'm at it.

The following is an excerpt

lazy
Defers loading the image until it reaches a calculated distance from the viewport, as defined by the browser. The intent is to avoid the network and storage bandwidth needed to handle the image until it's reasonably certain that it will be needed. This generally improves the performance of the content in most typical use cases.

Note: Loading is only deferred when JavaScript is enabled. This is an anti-tracking measure, because if a user agent supported lazy loading when scripting is disabled, it would still be possible for a site to track a user's approximate scroll position throughout a session, by strategically placing images in a page's markup such that a server can track how many images are requested and when.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img#attributes